# [IMP] history: reduce memory allocation