@robodoo 2023-11-16
[Pull request status dashboard](https://mergebot.odoo.com/odoo/o-spreadsheet/pull/3168).